- Secured with Jwt tokens

+++ src/main/kotlin/pl/integrable/dusterserver/service/JwtTokenPrincipal.kt
@@ -1,3 +1,3 @@
 package pl.integrable.dusterserver.service
 
-class JwtTokenPrincipal(val id: Long, val servicename: String, val role: String)
+class JwtTokenPrincipal(val id: Integer, val servicename: String, val role: String)
